backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<11.0
python-slugify>=8.0,<9.0
